requirements copy.txt
PyJWT~=1.7.1
PyYAML~=5.3.1
requests~=2.22.0
requests-unixsocket~=0.2.0
asgiref==3.4.1
click==8.0.1
cycler==0.10.0
fastapi==0.68.0
h11==0.12.0
kiwisolver==1.3.1
matplotlib==3.4.2
numpy==1.21.1
pandas==1.3.1
Pillow==8.3.1
pydantic==1.8.2
pyparsing==2.4.7
python-dateutil==2.8.2
pytz==2021.1
six==1.16.0
starlette==0.14.2
typing-extensions==3.10.0.0
uvicorn==0.14.0
python-dotenv~=0.17.1
PyYAML~=5.3.1
celery[redis]==4.0.2